Restructure project Fixes #177
- created folders under widgets folder and moved existing files
- that is literally the only changes in this commit
