+++ lib/widgets/chat_app.dart
@@ -1,12 +1,12 @@
-import 'package:adventures_in_chat_app/auth_page.dart';
-import 'package:adventures_in_chat_app/chat_page.dart';
+import 'package:adventures_in_chat_app/widgets/auth/auth_page.dart';
+import 'package:adventures_in_chat_app/widgets/messages/chat_page.dart';
 import 'package:adventures_in_chat_app/extensions/extensions.dart';
-import 'package:adventures_in_chat_app/home_page.dart';
-import 'package:adventures_in_chat_app/link_accounts_page.dart';
-import 'package:adventures_in_chat_app/profile_page.dart';
+import 'package:adventures_in_chat_app/widgets/home/home_page.dart';
+import 'package:adventures_in_chat_app/widgets/auth/link_accounts_page.dart';
+import 'package:adventures_in_chat_app/widgets/options/profile_page.dart';
 import 'package:adventures_in_chat_app/services/database_service.dart';
 import 'package:adventures_in_chat_app/services/fcm_service.dart';
-import 'package:adventures_in_chat_app/splash_page.dart';
+import 'package:adventures_in_chat_app/widgets/splash/splash_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
