+++ lib/widgets/conversations/user_search_page.dart
@@ -3,7 +3,7 @@ import 'dart:collection';
 import 'package:adventures_in_chat_app/extensions/extensions.dart';
 import 'package:adventures_in_chat_app/models/user_item.dart';
 import 'package:adventures_in_chat_app/services/database_service.dart';
-import 'package:adventures_in_chat_app/widgets/user_avatar.dart';
+import 'package:adventures_in_chat_app/widgets/shared/user_avatar.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
