+++ lib/widgets/messages/chat_page.dart
@@ -1,6 +1,6 @@
 import 'package:adventures_in_chat_app/models/message.dart';
 import 'package:adventures_in_chat_app/services/database_service.dart';
-import 'package:adventures_in_chat_app/widgets/chat_message.dart';
+import 'package:adventures_in_chat_app/widgets/messages/chat_message.dart';
 import 'package:flutter/material.dart';
 import 'package:adventures_in_chat_app/models/conversation_item.dart';
